+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Bugfix: Fix Ctrl + Backspace not closing colon emote picker. (#2780)
 - Bugfix: Approving/denying AutoMod messages works again. (#2779)
 - Dev: Migrated AutoMod approve/deny endpoints to Helix. (#2779)
+- Dev: Migrated Get Cheermotes endpoint to Helix. (#2440)
 
 ## 2.3.1